pyproject.toml
[project]
name = "markupdown"
version = "0.1.0"
description = "A modular static site generator in Python"
authors = [
    {name = "voidl00p", email = "null@voidl00p.net"},
]
dependencies = ["python-liquid>=1.12.2", "python-frontmatter>=1.1.0", "minify-html>=0.15.0", "mistune>=3.1.1"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "CC0-1.0"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["example", "src"]

[tool.pdm.scripts]
black = "black src/markupdown/ tests/"
isort = "isort src/markupdown/ tests/"
style = {composite = ["black", "isort"]}
test = "pytest"

[tool.isort]
profile = "black"
case_sensitive = true

[tool.pylint.master]
fail-under = 7

[tool.pylint.messages_control]
max-line-length = 120

[tool.pyright]
include = ["src/markupdown/"]
pythonPlatform = "All"

[dependency-groups]
test = [
    "pytest>=8.3.4",
]
style = [
    "black>=25.1.0",
    "isort>=5.12.0",
    "pylint>=2.17.5",
    "pyright>=1.1.329",
]
